backend/requirements.txt
alembic==1.17.2
annotated-types==0.7.0
anyio==4.11.0
bcrypt==5.0.0
black==25.11.0
certifi==2025.11.12
click==8.3.1
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hupper==1.12.1
idna==3.11
legacy-cgi==2.6.4
Mako==1.3.10
MarkupSafe==3.0.3
mypy_extensions==1.1.0
packaging==25.0
PasteDeploy==3.1.0
pathspec==0.12.1
plaster==1.1.2
plaster-pastedeploy==1.0.1
platformdirs==4.5.1
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
pyramid==2.0.2
pytokens==0.3.0
setuptools==80.9.0
sniffio==1.3.1
SQLAlchemy==2.0.44
translationstring==1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
venusian==3.1.1
waitress==3.0.2
WebOb==1.8.9
zope.deprecation==6.0
zope.interface==8.1.1
qrcode==7.4.2
pillow==10.1.0
pyzbar==0.1.9
